\docType{data}
\name{Abbas}
\alias{Abbas}
\title{Basis Matrix for Whole Blood Expression Deconvolution}
\source{
  Supplementary Table 1 of Abbas et al. (2009) Found at:
  doi:10.1371/journal.pone.0006098.s001 (0.11 MB PDF)
}
\description{
  Basis matrix for performing gene expression deconvolution
  of whole blood as in Abbas et al. (2009).
}
\details{
  The data was extracted from Supplementary Table 1 of
  Abbas et al. (2009), converted as text, and processed
  using the function \code{\link{.Abbas}}.
}
\references{
  Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark
  HF (2009). "Deconvolution of blood microarray data
  identifies cellular activation patterns in systemic lupus
  erythematosus." _PloS one_, *4*(7), pp. e6098. ISSN
  1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0006098>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/19568420>.
}
\seealso{
  Other Abbas: \code{\link{.Abbas}}, \code{\link{.mAbbas}},
  \code{\link{.SLE}}
}
\keyword{datasets}

