\docType{package}
\name{CellMix-package}
\alias{CellMix}
\alias{CellMix-package}
\title{Gene Expression Deconvolution}
\usage{
  CellMix()
}
\description{
  Gene Expression Deconvolution

  Prints/retrieve technical information or data about the
  \pkg{CellMix} package.
}
\details{
  \tabular{ll}{ Package: \tab CellMix\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2011-11-01\cr
  License: \tab GPL (>= 2)\cr LazyLoad: \tab yes\cr }
}
\examples{
# some technical info about the package
CellMix()
}
\author{
  Renaud Gaujoux \email{renaud@cbio.uct.ac.za}

  Maintainer: Renaud Gaujoux \email{renaud@cbio.uct.ac.za}
}
\keyword{package}

