\docType{methods}
\name{combine,MarkerList,factor-method}
\alias{combine,MarkerList,factor-method}
\title{Combine markers from multiple cell types of a MarkerList object, based on groups
defined by a given factor.}
\usage{
  \S4method{combine}{MarkerList,factor}(x, y, ...)
}
\arguments{
  \item{x}{One of the values.}

  \item{y}{A second value.}

  \item{...}{Any other objects of the same class as
  \code{x} and \code{y}.}
}
\description{
  Combine markers from multiple cell types of a MarkerList
  object, based on groups defined by a given factor.
}
\keyword{methods}

