\docType{methods}
\name{Compare,MarkerList,numeric-method}
\alias{Compare,MarkerList,numeric-method}
\title{Operations on MarkerList Objects}
\usage{
  \S4method{Compare}{MarkerList,numeric}(e1, e2)
}
\arguments{
  \item{e1}{objects.}

  \item{e2}{objects.}
}
\description{
  \code{Compare} compares all embedded values with a given
  fixed value. This is useful to filter markers based on
  their associated scores.
}
\keyword{methods}

