\name{.Abbas}
\alias{.Abbas}
\title{Processing for Whole Blood Deconvolution Basis Matrix}
\usage{
  .Abbas(file = "TableS1.txt")
}
\arguments{
  \item{file}{input file. The file would have been created
  from the conversion into text of the Supplementary Table
  S1 (a strange PDF file) from Abbas et al. (2009).}
}
\description{
  Function used to create the deconvolution basis matrix
  \code{\link{Abbas}}, from supplementary data of Abbas et
  al. (2009).
}
\references{
  Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark
  HF (2009). "Deconvolution of blood microarray data
  identifies cellular activation patterns in systemic lupus
  erythematosus." _PloS one_, *4*(7), pp. e6098. ISSN
  1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0006098>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/19568420>.
}
\seealso{
  Other Abbas: \code{\link{.mAbbas}}, \code{\link{.SLE}}
}
\keyword{internal}

