\docType{methods}
\name{.DollarNames,ExpressionMix-method}
\alias{.DollarNames,ExpressionMix-method}
\title{Auto-completion for \code{\linkS4class{ExpressionMix}} objects}
\usage{
  \S4method{.DollarNames}{ExpressionMix}(x, pattern = "")
}
\arguments{
  \item{x}{ An R object for which valid names after
  \code{"$"} are computed and returned.  }

  \item{pattern}{ A regular expression.  Only matching
  names are returned.  }
}
\description{
  Auto-completion for \code{\linkS4class{ExpressionMix}}
  objects
}
\keyword{methods}

