\name{.IRIS}
\alias{.IRIS}
\title{Processing for the Immune Response In-Silico (IRIS) Database}
\usage{
  .IRIS(file = "IRIS.txt")
}
\arguments{
  \item{file}{input file. This should be a tab separated
  version of the Supplementary Table 1 from Abbas et al.
  (2005).}
}
\description{
  Function used to create the \code{\link{IRIS}} data (an
  \code{link{ExpressionSet}} object).
}
\references{
  Abbas AR, Baldwin D, Ma Y, Ouyang W, Gurney A, Martin F,
  Fong S, van Lookeren Campagne M, Godowski P, Williams PM,
  Chan aC and Clark HF (2005). "Immune response in silico
  (IRIS): immune-specific genes identified from a
  compendium of microarray expression data." _Genes and
  immunity_, *6*(4), pp. 319-31. ISSN 1466-4879, <URL:
  http://dx.doi.org/10.1038/sj.gene.6364173>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/15789058>.
}
\seealso{
  Other IRIS: \code{\link{.mIRIS}}
}
\keyword{internal}

