\name{.mAbbas}
\alias{Abbas-markers}
\alias{.mAbbas}
\title{Optimised Set of Marker Genes for Immune Cells}
\source{
  Supplementary Table 1 of Abbas et al. (2009) Found at:
  doi:10.1371/journal.pone.0006098.s001 (0.11 MB PDF)
}
\usage{
  .mAbbas(sparsity = 0.8,
    id = c("AFFY", "ENTREZID", "SYMBOL"))
}
\arguments{
  \item{sparsity}{sparsity threshold (see
  \code{\link{sparseness}})}

  \item{id}{original identifier to use.}
}
\description{
  \pkg{CellMix} access key: \dQuote{Abbas}

  Curated list of marker probesets for immune cell types
  from the \code{\link{Abbas}} data (\cite{Abbas et al.
  (2009)}). It contains Affymetrix probesets of markers for
  B cells, dendritic cells, monocytes, neutrophils, NK
  cells and T cells, as well as for the lymphoid and
  myeloid lineages, and the general category of all these
  immune cells.
}
\details{
  The \code{MarkerList} object is generated by the internal
  function \code{.mAbbas}, to which arguments can be passed
  via \code{MarkerList('Abbas', ...)} to build lists, using
  different original identifiers and filtering threshold.

  However, it is recommended to rather use the function
  \code{\link{convertIDs}} to convert the primary
  identifiers into other identifiers, using up-to-date
  annotation packages.

  The markers are ordered in each cell type by decreasing
  sparsity value, i.e. from the more to the less cell
  type-specific, computed from the optimised signature
  matrix \code{\link{Abbas}}.
}
\examples{
MarkerList('Abbas')
}
\references{
  Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark
  HF (2009). "Deconvolution of blood microarray data
  identifies cellular activation patterns in systemic lupus
  erythematosus." _PloS one_, *4*(7), pp. e6098. ISSN
  1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0006098>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/19568420>.
}
\seealso{
  \code{\link{IRIS-markers}}

  Other Abbas: \code{\link{.Abbas}}, \code{\link{.SLE}}
}

