\name{.mHaemAtlas}
\alias{HaemAtlas-markers}
\alias{.mHaemAtlas}
\title{HaemAtlas Marker List for Immune Cells - Watkins et al. (2009)}
\usage{
  .mHaemAtlas(id = c("ILMN", "nuID", "ENTREZID", "SYMBOL", "_ILMN", "_ENSEMBL", "_SYMBOL"))
}
\arguments{
  \item{id}{original identifier to use.}
}
\description{
  \pkg{CellMix} access key: \dQuote{HaemAtlas}

  Marker list created from the \code{\link{HaemAtlas}}
  dataset (\cite{Watkins et al. (2009)}). It contains
  markers for CD4+ and CD8+ T-cells, lymphocytes,
  monocytes, B-cells, NK cells, and granulocytes.
}
\details{
  The \code{MarkerList} object is generated by the internal
  function \code{.mHaemAtlas}, to which arguments can be
  passed via \code{MarkerList('HaemAtlas', ...)} to build
  lists using different identifiers.

  However, it is recommended to rather use the function
  \code{\link{convertIDs}} to convert the primary
  identifiers into other identifiers, using up-to-date
  annotation packages.
}
\examples{
MarkerList('HaemAtlas')
}
\references{
  Watkins Na, Gusnanto A, Bono Bd, De S, Miranda-Saavedra
  D, Hardie DL, Angenent WGJ, Attwood AP, Ellis PD, Erber
  W, Foad NS, Garner SF, Isacke CM, Jolley J, Koch K,
  Macaulay IC, Morley SL, Rendon A, Rice KM, Taylor N,
  Thijssen-Timmer DC, Tijssen MR, Schoot CEvd, Wernisch L,
  Winzer T, Dudbridge F, Buckley CD, Langford CF, Teichmann
  S, Gottgens B and Ouwehand WH (2009). "A HaemAtlas:
  characterizing gene expression in differentiated human
  blood cells." _Blood_, *113*(19), pp. e1-9. ISSN
  1528-0020, <URL:
  http://dx.doi.org/10.1182/blood-2008-06-162958>, <URL:
  http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2680378\&tool=pmcentrez\&rendertype=abstract>.
}

