\name{.mIRIS}
\alias{IRIS-markers}
\alias{.mIRIS}
\title{Marker Genes for Immune Cells (IRIS)}
\source{
  \url{http://share.gene.com/clark.iris.2004/iris/iris.html}
}
\usage{
  .mIRIS(id = c("AFFY", "ENTREZID", "ENSEMBL"),
    score = "F", threshold = 0)
}
\arguments{
  \item{id}{original identifier to use.}

  \item{score}{name of the column from the
  \code{\link{IRIS}} dataset to use as a score.}

  \item{threshold}{minimum value threshold for filtering on
  scores.}
}
\description{
  \pkg{CellMix} access key: \dQuote{IRIS}

  List of marker probesets for immune cell types from the
  \code{\link{IRIS}} dataset (\cite{Abbas et al. (2005)}).
  It contains Affymetrix probesets of markers for B cells,
  dendritic cells, monocytes, neutrophils, NK cells and T
  cells, as well as for the lymphoid and myeloid lineages,
  and the general category of all these immune cells.
}
\details{
  The \code{MarkerList} object is generated by the internal
  function \code{.mIRIS}, to which arguments can be passed
  via \code{MarkerList('IRIS', ...)} to build lists, using
  different original identifiers and filtering threshold.

  However, it is recommended to rather use the function
  \code{\link{convertIDs}} to convert the primary
  identifiers into other identifiers, using up-to-date
  annotation packages.

  The markers are ordered in each cell type by decreasing F
  value (from \code{fData(IRIS)$F}), i.e. from the more to
  the less cell type-specific.
}
\examples{
MarkerList('IRIS')
}
\references{
  Abbas AR, Baldwin D, Ma Y, Ouyang W, Gurney A, Martin F,
  Fong S, van Lookeren Campagne M, Godowski P, Williams PM,
  Chan aC and Clark HF (2005). "Immune response in silico
  (IRIS): immune-specific genes identified from a
  compendium of microarray expression data." _Genes and
  immunity_, *6*(4), pp. 319-31. ISSN 1466-4879, <URL:
  http://dx.doi.org/10.1038/sj.gene.6364173>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/15789058>.
}
\seealso{
  Other IRIS: \code{\link{.IRIS}}
}

