\name{.mPalmer}
\alias{.mPalmer}
\alias{Palmer-markers}
\title{Marker List for Immune Cells - Palmer et al. (2006)}
\usage{
  .mPalmer(id = c("SYMBOL", "LUID"),
    score = c("Corr", "Fold"), threshold = 0)
}
\arguments{
  \item{id}{original identifier to use.}

  \item{score}{name of the column from the
  \code{\link{IRIS}} dataset to use as a score.}

  \item{threhold}{minimum value threshold for filtering on
  scores.}
}
\description{
  \pkg{CellMix} access key: \dQuote{Palmer}

  Marker list created from the \code{Palmer} dataset
  (\cite{Palmer et al. (2006)}). It contains markers for
  B-cells, CD4+ and CD8+ T-cells, lymphocytes and
  granulocytes.
}
\details{
  The \code{MarkerList} object is generated by the internal
  function \code{.mPalmer}, to which arguments can be
  passed via \code{MarkerList('Palmer', ...)} to build
  lists using different identifiers and/or filtering
  threshold.

  However, it is recommended to rather use the function
  \code{\link{convertIDs}} to convert the primary
  identifiers into other identifiers, using up-to-date
  annotation packages.
}
\examples{
MarkerList('Palmer')
}
\references{
  Palmer C, Diehn M, Alizadeh Aa and Brown PO (2006).
  "Cell-type specific gene expression profiles of
  leukocytes in human peripheral blood." _BMC genomics_,
  *7*, pp. 115. ISSN 1471-2164, <URL:
  http://dx.doi.org/10.1186/1471-2164-7-115>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/16704732>.
}
\keyword{internal}

