\name{.mTDDB_HS}
\alias{.mTDDB_HS}
\alias{TDDB_HS}
\alias{TissueDistributionDB_HS-markers}
\title{Human Tissue Specific Genes from the TissueDistributionDB Database}
\usage{
  .mTDDB_HS(percent = 99)
}
\arguments{
  \item{percent}{minimum percentage expression value
  required for a gene to be included as a marker.}
}
\description{
  \pkg{CellMix} access key:
  \dQuote{TissueDistributionDB_HS}

  Marker list for Human tissues based on Unigene EST
  clusters, downloaded from the TissueDistributionDB
  database (\cite{Kogenaru et al. (2009)}).
}
\details{
  The \code{MarkerList} object is generated by the internal
  function \code{.mTDDB_HS}, to which arguments can be
  passed via \code{MarkerList('TDDB_HS', ...)} to build
  lists using a different filtering threshold.
}
\examples{
MarkerList('TDDB_HS')
}
\references{
  Kogenaru S, Val C, Hotz-Wagenblatt A and Glatting K
  (2009). "TissueDistributionDBs: a repository of
  organism-specific tissue-distribution profiles."
  _Theoretical Chemistry Accounts_, *125*(3-6), pp.
  651-658. ISSN 1432-881X, <URL:
  http://dx.doi.org/10.1007/s00214-009-0670-5>, <URL:
  http://www.springerlink.com/index/10.1007/s00214-009-0670-5
  http://www.springerlink.com/index/V03177J8214R3341.pdf>.
}
\seealso{
  Other TDDB: \code{\link{.mTDDB_RN}}
}

