\name{.mTIGER}
\alias{.mTIGER}
\alias{TIGER-markers}
\title{TiGER - Human Tissue Specific Genes}
\usage{
  .mTIGER(score = 5, log10P = 3.47853)
}
\arguments{
  \item{score}{minimum score for a gene to be included in
  the list}

  \item{log10P}{minimum value for -log10(P-value) for a
  gene to be included in the list}
}
\description{
  \pkg{CellMix} access key: \dQuote{TIGER}

  Marker list created from the \code{\link{TIGER}} dataset
  (\cite{Liu et al. (2008)}). The TiGER database is based
  on UniGene EST distribution profiles. The markers are
  defined by thresholds on enrichment scores and their
  associated log10 p-values.
}
\details{
  The \code{MarkerList} object is generated by the internal
  function \code{.mTIGER}, to which arguments can be passed
  via \code{MarkerList('TIGER', ...)} to build lists using
  different filtering threshold.

  The function \code{\link{convertIDs}} can be used to
  convert the primary identifiers into other identifiers
  (e.g. ENTREZ ids or microarray probe ids), using
  up-to-date mappings from Bioconductor annotation
  packages.
}
\examples{
MarkerList('TIGER')
}
\references{
  Liu X, Yu X, Zack DJ, Zhu H and Qian J (2008). "TiGER: a
  database for tissue-specific gene expression and
  regulation." _BMC bioinformatics_, *9*, pp. 271. ISSN
  1471-2105, <URL:
  http://dx.doi.org/10.1186/1471-2105-9-271>, <URL:
  http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2438328\&tool=pmcentrez\&rendertype=abstract>.
}

