\name{.SLE}
\alias{.SLE}
\title{Processing for SLE Data}
\usage{
  .SLE(file = "TableS2.txt")
}
\arguments{
  \item{file}{input file. The file would be the XLS
  (actually tab separated) file provided as Supplementary
  Table S2 from Abbas et al. (2009).}
}
\value{
  an \code{\link{ExpressionSet}} object
}
\description{
  Function used to create the \code{\link{SLE}} data.
}
\references{
  Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark
  HF (2009). "Deconvolution of blood microarray data
  identifies cellular activation patterns in systemic lupus
  erythematosus." _PloS one_, *4*(7), pp. e6098. ISSN
  1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0006098>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/19568420>.
}
\seealso{
  Other Abbas: \code{\link{.Abbas}}, \code{\link{.mAbbas}}
}
\keyword{internal}

