\docType{methods}
\name{eset}
\alias{eset}
\alias{eset,ExpressionMix-method}
\alias{eset,GEDdata_entry-method}
\alias{eset-methods}
\title{Extracting Expression Data}
\usage{
  eset(object, ...)

  \S4method{eset}{GEDdata_entry}(object, load = TRUE,
    filter = TRUE, verbose = FALSE, save = TRUE,
    cache = TRUE)
}
\arguments{
  \item{object}{object from which to extract the expression
  data}

  \item{...}{extra argument to allow extension. See each
  method's description for more details,.}

  \item{load}{logical that indicates if the primary data
  should be (re-)loaded, or a character string giving the
  path of the file from where to load the data, which will
  typically be the Matrix file as downloaded from GEO. If
  \code{FALSE} the gene expression data is not loaded and
  the result will be \code{NULL} if the dataset has not
  already been loaded in a previous call.

  If explicitly \code{TRUE} (i.e. not missing), then the
  data is (re-)loaded from cache, or downloaded from a
  public repository if it has not been previously cached.
  In this case, the pre-processing pipeline is performed
  again.}

  \item{filter}{logical that indicates if the preprocessing
  pipeline should be applied to the "plain" data. If
  \code{TRUE} then the processing/filtering function
  associated with the data (registry field \code{'filter'})
  is applied. A custom pre-processing function may also be
  passed.

  If \code{FALSE}, no processing is performed on the data.}

  \item{verbose}{logical/number that toggles verbose
  messages. To better monitor how the data is
  generated/processed, it is recommended to toggle on
  verbosity, and use numeric values at various levels (e.g.
  \code{verbose=3}) to display more detailed outputs.}

  \item{save}{logical that indicates if the data should be
  saved in in R binary format in the local GED cache, for
  fast future loading -- possibly in a different R
  session.}

  \item{cache}{logical that indicates if the data should be
  retrieved from the local cache if present, or
  re-(down)loaded. If some remote data has been downloaded
  by a previous call to \code{eset}, these data are
  generally loaded from disk and not re-downloaded by the
  relevant package (e.g., \pkg{GEOquery} or
  \pkg{ArrayExpress}). If \code{save=TRUE} (as default),
  the new data will replace any previously downloaded
  data.}
}
\value{
  an \code{\link{ExpressionSet}} object.
}
\description{
  Extracts expression data (usually as \code{ExpressionSet}
  objects) from other objects.
}
\section{Methods}{
  \describe{

  \item{eset}{\code{signature(object = "ExpressionMix")}:
  Extracts an ExpressionSet object from an
  \code{ExpressionMix} object.

  Because \code{\linkS4class{ExpressionMix}} inherits from
  \code{\linkS4class{ExpressionSet}}, this methods simply
  returns the object coerced as an \code{ExpressionSet}
  object. }

  \item{eset}{\code{signature(object = "GEDdata_entry")}:
  Extracting Expression Data from CellMix Dataset Registry
  Entry.

  Loads and extracts the expression data from a CellMix
  registry dataset entry, as an \code{\link{ExpressionSet}}
  object.

  The first time the expression data is accessed, and if
  argument \code{load=TRUE}, it is loaded from the local
  storage if possible (cf. \code{\link{GEDsave}}), or
  downloaded from GEO using \code{\link{GEDdownload}}. It
  is then pre-processed (see below) and the resulting value
  (an \code{ExpressionSet} object) is cached in the
  GEDdata_entry environment. Subsequent calls use the
  cached value.

  The data pre-processing consists in filtering and
  extraction/addition of phenotypic or feature annotation
  data. The filtering is performed by the function stored
  in element \code{object$filter}. The extra phenotypic
  data is generated by the function stored in element
  \code{object$pdata}. The extra feature data is generated
  by the function stored in element \code{object$fdata}.
  Note that the whole pre-processing pipeline is not
  cached. }

  }
}
\examples{
\dontrun{
# get data from the registry
e <- gedData('GSE20300')
# this is an ExpressionMix object
e
# extract the ExpressionSet part
class(eset(e))
}
}
\seealso{
  \code{\link{ExpressionSet}}
}
\keyword{methods}

