\docType{class}
\name{GEDStrategy-class}
\alias{GEDStrategy-class}
\alias{show,GEDStrategy-method}
\title{Class for GED Algorithms}
\usage{
  \S4method{show}{GEDStrategy}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  Class for GED Algorithms
}
\section{Methods}{
  \describe{

  \item{GEDStrategy}{\code{signature(object =
  "GEDStrategy")}: Constructor-Copy method, which creates
  an \code{GEDalgorithm} object from a template and ensure
  that the object has a key that does not conflict with
  registered algorithms. }

  }
}
\keyword{methods}

