\docType{data}
\name{GPL2bioc}
\alias{GPL2bioc}
\title{Map Between Bioconductor Annotation Packages and GEO GPL Identifiers}
\description{
  The data \code{GPL2bioc} contains a data.frame that was
  built using the function \code{getBiocPlatformMap} from
  the \pkg{GEOmetadb} package.
}
\keyword{datasets}

