\docType{data}
\name{GSE11058_pdata}
\alias{GSE11058_pdata}
\title{Cell Line Proportions for Dataset GSE11058}
\source{
  \url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE11058}
}
\description{
  The proportions were extracted from the GEO page for
  dataset GSE11058.
}
\references{
  Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark
  HF (2009). "Deconvolution of blood microarray data
  identifies cellular activation patterns in systemic lupus
  erythematosus." _PloS one_, *4*(7), pp. e6098. ISSN
  1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0006098>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/19568420>.
}
\seealso{
  Other Abbas: \code{\link{.Abbas}}, \code{\link{.mAbbas}},
  \code{\link{.SLE}}
}
\keyword{datasets}

