\docType{data}
\name{GSE20300_pdata}
\alias{GSE20300_pdata}
\title{Complete Blood Count for Dataset GSE20300}
\source{
  Extracted from Supplementary Table 3 of Shen-Orr et al.
  (2010)
}
\description{
  The data was extracted from the PDF file and reordered in
  the same order as the microarray dataset GSE20300,
  assuming that the GEO samples' order within each group
  (Stable or ACR) is consistent with the samples' order in
  the PDF file.
}
\references{
  Shen-Orr SS, Tibshirani R, Khatri P, Bodian DL, Staedtler
  F, Perry NM, Hastie T, Sarwal MM, Davis MM and Butte AJ
  (2010). "Cell type-specific gene expression differences
  in complex tissues." _Nature methods_, *7*(4), pp. 287-9.
  ISSN 1548-7105, <URL:
  http://dx.doi.org/10.1038/nmeth.1439>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/20208531>.
}
\keyword{datasets}

