\name{GSE24223_fdata}
\alias{.GSE24223_fdata}
\alias{GSE24223_fdata}
\title{Dataset GSE24223: Cytometry Antigen Gene Expression}
\source{
  doi:10.1371/journal.pone.0013358.s007 (0.14 MB XLS)
  \url{http://www.plosone.org/article/fetchSingleRepresentation.action?uri=info:doi/10.1371/journal.pone.0013358.s007}

  doi:10.1371/journal.pone.0013358.s007 (0.14 MB XLS)
  \url{http://www.plosone.org/article/fetchSingleRepresentation.action?uri=info:doi/10.1371/journal.pone.0013358.s007}
}
\usage{
  .GSE24223_fdata(file = "journal.pone.0013358.s007.xls")
}
\arguments{
  \item{file}{source file}
}
\description{
  Supplemental Table 5 from \cite{Grigoryev et al. (2010)}
  containing average gene expression values for cytometry
  antigen markers, within each group of samples. Groups are
  defined by the pair of experimental factors (Type, Time)
  (i.e. Cell type x Time of blood draw).

  \code{.GSE24223_fdata} is the function used to create the
  \code{\link{GSE24223_fdata}} data.
}
\references{
  Grigoryev Ya, Kurian SM, Avnur Z, Borie D, Deng J,
  Campbell D, Sung J, Nikolcheva T, Quinn A, Schulman H,
  Peng SL, Schaffer R, Fisher J, Mondala T, Head S,
  Flechner SM, Kantor AB, Marsh C and Salomon DR (2010).
  "Deconvoluting post-transplant immunity: cell
  subset-specific mapping reveals pathways for activation
  and expansion of memory T, monocytes and B cells." _PloS
  one_, *5*(10), pp. e13358. ISSN 1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0013358>, <URL:
  http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2954794\&tool=pmcentrez\&rendertype=abstract>.
}
\keyword{internal}

