\docType{data}
\name{GSE3649_pdata}
\alias{GSE3649_pdata}
\title{Phenotypic Annotation Data for Dataset GSE3649}
\source{
  \url{http://genome-www.stanford.edu/normalblood/}
}
\description{
  This data provides several phenotypic information about
  the samples contained in the GEO dataset GSE3649, like
  age, gender, time of blood draw, as well as clinical data
  such as total white count, differential counts for
  neutrophils, lymphocytes, monocytes, eosinophils, and
  basophils, red blood cell count, platelet count,
  hemoglobin, hematocrit, and erythrocyte indices [mean
  corpuscular volume, mean corpuscular hemoglobin, mean
  corpuscular hemoglobin concentration, and red cell
  distribution width (RDW)].
}
\details{
  Although mentioned in Whitney et al. (2003), these data
  were not initially available online. We thank Stephen
  Popper who made them available on our request.
}
\note{
  Around November 2011, we experienced issues in accessing
  the web page.
}
\references{
  Whitney A, Diehn M, Popper S, Alizadeh A, Boldrick J,
  Relman D and Brown P (2003). "Individuality and variation
  in gene expression patterns in human blood." _Proceedings
  of the National Academy of Sciences of the United States
  of America_, *100*(4), pp. 1896. <URL:
  http://www.pnas.org/content/100/4/1896.short>.
}
\seealso{
  Other Whitney: \code{\link{.GSE3649_convert}},
  \code{\link{.GSE3649_pdata}}
}
\keyword{datasets}

