\docType{methods}
\name{initialize,ExpressionMix-method}
\alias{initialize,ExpressionMix-method}
\title{Initialize method for ExpressionMix object}
\usage{
  \S4method{initialize}{ExpressionMix}(.Object, ...)
}
\arguments{
  \item{.Object}{ An object: see the Details section.}

  \item{...}{data to include in the new object.  Named
  arguments correspond to slots in the class definition.
  Unnamed arguments must be objects from classes that this
  class extends.}
}
\description{
  Initialize method for ExpressionMix object
}
\keyword{methods}

