\docType{data}
\name{IRIS}
\alias{IRIS}
\title{Immune Response In-Silico (IRIS) Data}
\source{
  \url{http://share.gene.com/clark.iris.2004/irisSup/Sup6.xls}
}
\description{
  Extracted from Supplementary Table 1 from Abbas et al.
  (2005).
}
\references{
  Abbas AR, Baldwin D, Ma Y, Ouyang W, Gurney A, Martin F,
  Fong S, van Lookeren Campagne M, Godowski P, Williams PM,
  Chan aC and Clark HF (2005). "Immune response in silico
  (IRIS): immune-specific genes identified from a
  compendium of microarray expression data." _Genes and
  immunity_, *6*(4), pp. 319-31. ISSN 1466-4879, <URL:
  http://dx.doi.org/10.1038/sj.gene.6364173>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/15789058>.
}
\seealso{
  Other IRIS: \code{\link{.IRIS}}, \code{\link{.mIRIS}}
}
\keyword{datasets}

