\name{MarkerSetCollection}
\alias{MarkerSetCollection}
\alias{MarkerSetCollection-methods}
\title{Factory Methods for Marker Gene Set Collections}
\usage{
  MarkerSetCollection(object, ...)
}
\arguments{
  \item{object}{source object from which the object must be
  created}

  \item{...}{extra arguments to allow extension and passed
  to subsequent calls. See details in each method's
  description.}
}
\description{
  Factory Methods for Marker Gene Set Collections
}

