\docType{data}
\name{markersGrigoryev}
\alias{markersGrigoryev}
\title{Cytometry Antigen Expression from Grigoryev et al. (2010)}
\source{
  \url{http://www.plosone.org/article/fetchSingleRepresentation.action?uri=info:doi/10.1371/journal.pone.0013358.s007}
}
\description{
  This dataset contains probe ids of cytometry antigen
  marker genes for immune cells, and their expression on
  Affymetrix HGU 133 Plus 2.0 chips, generated according to
  the experimental design from the study in \cite{Grigoryev
  et al. (2010)}.
}
\details{
  There are markers for B cells, Neutrophils, T cells,
  Monocytes, NK cells, Platelets, Eosinophils, as well as
  for whole blood cells as a generic type, and some
  combination of cell types, e.g., Monocyte, B cell.
}
\references{
  Grigoryev Ya, Kurian SM, Avnur Z, Borie D, Deng J,
  Campbell D, Sung J, Nikolcheva T, Quinn A, Schulman H,
  Peng SL, Schaffer R, Fisher J, Mondala T, Head S,
  Flechner SM, Kantor AB, Marsh C and Salomon DR (2010).
  "Deconvoluting post-transplant immunity: cell
  subset-specific mapping reveals pathways for activation
  and expansion of memory T, monocytes and B cells." _PloS
  one_, *5*(10), pp. e13358. ISSN 1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0013358>, <URL:
  http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2954794\&tool=pmcentrez\&rendertype=abstract>.
}
\keyword{datasets}

