\docType{data}
\name{Palmer}
\alias{Palmer}
\title{Palmer Dataset}
\source{
  \url{http://www.biomedcentral.com/content/supplementary/1471-2164-7-115-S2.xls}
}
\description{
  Markers for B-cells, CD4+ and CD8+ T-cells, lymphocytes
  and granulocytes from \cite{Palmer et al. (2006)}. The
  dataset was created using the internal function
  \code{.Palmer}.
}
\references{
  Palmer C, Diehn M, Alizadeh Aa and Brown PO (2006).
  "Cell-type specific gene expression profiles of
  leukocytes in human peripheral blood." _BMC genomics_,
  *7*, pp. 115. ISSN 1471-2164, <URL:
  http://dx.doi.org/10.1186/1471-2164-7-115>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/16704732>.
}
\seealso{
  Other Palmer: \code{\link{.mPalmer}}
}
\keyword{datasets}

