\name{regcheck}
\alias{checkS4}
\alias{regcheck}
\title{Validity Functions of Registry Fields}
\usage{
  checkS4(x)
}
\arguments{
  \item{x}{value of the field to be validated}
}
\description{
  Validity Functions of Registry Fields

  Checks if \code{x} is a valid S4 object by calling
  \code{\link{validObject}} on it.
}

