\name{requireFile}
\alias{requireFile}
\title{Require Existing File}
\usage{
  requireFile(file, msg = NULL, sep = " - ")
}
\arguments{
  \item{file}{File path to check.}

  \item{msg}{message to prepend to the error message.}

  \item{sep}{separator between \code{msg} and general error
  message.}
}
\description{
  Utility function to require a file to exist.
}
\keyword{internal}

