\docType{data}
\name{SLE}
\alias{SLE}
\title{Gene Expression from Systemic Lupus Erythematosus Patients}
\source{
  Extracted from Supplementary Table 2 of Abbas et al.
  (2009) Found at: doi:10.1371/journal.pone.0006098.s002
  (0.31 MB XLS)
}
\description{
  Expression data for the basis probesets for the main
  clinical test cohort of 118 SLE patients and healthy
  controls used in Abbas et al. (2009).
}
\references{
  Abbas AR, Wolslegel K, Seshasayee D, Modrusan Z and Clark
  HF (2009). "Deconvolution of blood microarray data
  identifies cellular activation patterns in systemic lupus
  erythematosus." _PloS one_, *4*(7), pp. e6098. ISSN
  1932-6203, <URL:
  http://dx.doi.org/10.1371/journal.pone.0006098>, <URL:
  http://www.ncbi.nlm.nih.gov/pubmed/19568420>.
}
\seealso{
  Other Abbas: \code{\link{.Abbas}}, \code{\link{.mAbbas}},
  \code{\link{.SLE}}
}
\keyword{datasets}

