\docType{data}
\name{TIGER}
\alias{TIGER}
\title{TiGER Database}
\source{
  \url{http://bioinfo.wilmer.jhu.edu/tiger}
  \url{http://bioinfo.wilmer.jhu.edu/tiger/download/tss_spf_rsc.txt}
  \url{http://bioinfo.wilmer.jhu.edu/tiger/download/hs2tissue-Table.txt}
}
\description{
  TiGER is a database developed by the Bioinformatics Lab
  at Wilmer Eye Institute of Johns Hopkins University. The
  database contains tissue-specific gene expression
  profiles or expressed sequence tag (EST) data,
  cis-regulatory module (CRM) data, and combinatorial gene
  regulation data.
}
\details{
  The dataset was created using the internal function
  \code{.TIGER}.
}
\references{
  Liu X, Yu X, Zack DJ, Zhu H and Qian J (2008). "TiGER: a
  database for tissue-specific gene expression and
  regulation." _BMC bioinformatics_, *9*, pp. 271. ISSN
  1471-2105, <URL:
  http://dx.doi.org/10.1186/1471-2105-9-271>, <URL:
  http://www.pubmedcentral.nih.gov/articlerender.fcgi?artid=2438328\&tool=pmcentrez\&rendertype=abstract>.
}
\seealso{
  Other TIGER: \code{\link{.mTIGER}}
}
\keyword{datasets}

