\docType{data}
\name{TissueDistributionDB_HS}
\alias{TissueDistributionDB_HS}
\title{Human Tissue Specific Genes from TissueDistributionDB}
\description{
  Dataset \code{TissueDistributionDB_HS} contains the
  complete original data obtained from
  \emph{TissueDistributionDB}. It was processed using the
  internal function \code{.TissueDistributionDB_HS}.
}
\references{
  Kogenaru S, Val C, Hotz-Wagenblatt A and Glatting K
  (2009). "TissueDistributionDBs: a repository of
  organism-specific tissue-distribution profiles."
  _Theoretical Chemistry Accounts_, *125*(3-6), pp.
  651-658. ISSN 1432-881X, <URL:
  http://dx.doi.org/10.1007/s00214-009-0670-5>, <URL:
  http://www.springerlink.com/index/10.1007/s00214-009-0670-5
  http://www.springerlink.com/index/V03177J8214R3341.pdf>.
}
\seealso{
  Other TDDB: \code{\link{.mTDDB_HS}},
  \code{\link{.mTDDB_RN}}
}
\keyword{datasets}

