\docType{data}
\name{VeryGene}
\alias{VeryGene}
\title{VeryGene Dataset}
\source{
  \url{http://www.verygene.com}
  \url{http://www.verygene.com/database.rar}
}
\description{
  VeryGene is developed as a curated, web-accessible
  centralized database for the annotation of
  tissue-specific/enriched genes (\cite{Yang et al.
  (2011)}).
}
\details{
  It currently contains entries for 3960 human genes
  covering 128 normal tissue/cell types compiled from the
  expression profiling of two large microarray data sets
  from Liang et al. (2006) and Su et al. (2004).

  It brings together much-needed information on preferred
  tissue/subcellular localization, functional annotation,
  pathway, mammalian phenotype, related diseases and
  targeting drug associated with any of these genes as a
  result of data integration from multiple sources.
  Information can be searched through gene, tissue and
  disease views and search result can be downloaded easily.

  The dataset was created with the internal functions
  \code{.VeryGene} and \code{MySQLtoSQLite}, to import the
  data from the MySQL dump available online.
}
\references{
  Yang X, Ye Y, Wang G, Huang H, Yu D and Liang S (2011).
  "VeryGene: linking tissue-specific genes to diseases,
  drugs, and beyond for knowledge discovery."
  _Physiological genomics_, *43*(8), pp. 457-60. ISSN
  1531-2267, <URL:
  http://dx.doi.org/10.1152/physiolgenomics.00178.2010>,
  <URL:
  http://physiolgenomics.physiology.org/cgi/content/abstract/43/8/457>.
}
\keyword{datasets}

