\name{VeryGene-markers}
\alias{VeryGene-markers}
\title{VeryGene - Marker List for Human Tissues}
\source{
  \url{http://www.verygene.com}
}
\description{
  Marker gene list for Human tissues created from the
  \code{\link{VeryGene}} dataset. Genes are ordered
  according to their specificity index.
}
\examples{
MarkerList('VeryGene')
}

